+++ src/main/java/org/sopt/like/service/LikeService.java
@@ -1,14 +1,14 @@
-package org.sopt.service;
+package org.sopt.like.service;
 
-import org.sopt.domain.Like;
-import org.sopt.domain.Post;
-import org.sopt.domain.User;
+import org.sopt.like.entity.Like;
+import org.sopt.post.entity.Post;
+import org.sopt.user.entity.User;
 import org.sopt.global.code.ErrorCode;
 import org.sopt.global.exception.BaseException;
 import org.sopt.global.exception.PostNotFoundException;
-import org.sopt.repository.LikeRepository;
-import org.sopt.repository.PostRepository;
-import org.sopt.repository.UserRepository;
+import org.sopt.like.repository.LikeRepository;
+import org.sopt.post.repository.PostRepository;
+import org.sopt.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
